Split config_type into config_map and from_file for easier understanding
